fix update bloom filter while resyncing
