+++ test/integration/resync.test.js
@@ -9,16 +9,10 @@ test.before(setup)
 test.after.always(close)
 
 /*
-    Scenario part 1: the wallet has created more than 20 used addresses (the pre-generated ones when we init the wallet).
-    Then we need to update the bloom filter with more addresses and see if we get the transactions for the new addresses.
-
-    Scenario part 2: We delete the database and try to resync. We should have all the transactions.
+    Scenario: generate 40 transactions with 40 addresses, erase wallet database and resync from mnemonic.
 */
 
-// new element to add to filter
-var newHash
-
-test.serial('should generate 20 transactions', async t => {
+test.serial('should generate 40 transactions', async t => {
   const { wallet, container } = t.context
   const containerName = (await container.inspect()).Name
 
@@ -26,20 +20,22 @@ test.serial('should generate 20 transactions', async t => {
   execSync(`docker exec ${containerName} dogecoin-cli generate 150`)
 
   // use 20 addresses
-  for (let i = 0; i < 20; i++) {
-    const address = await wallet.generateAddress()
+  for (let i = 0; i < 40; i++) {
+    const address = await wallet.getAddress()
 
     t.log(`Sending to address : ${address}`)
 
     const txid = execSync(`docker exec ${containerName} dogecoin-cli sendtoaddress ${address} 150`)
     execSync(`docker exec ${containerName} dogecoin-cli generate 50`)
+
+    const pubkeyHash = bs58check.decode(address).slice(1)
+    await wallet.markPubkeyAsUsed(pubkeyHash)
   }
 
   t.pass()
 })
 
-
-test.serial('should sync the 20 transcations', async t => {
+test.serial('should sync all the transactions', async t => {
   const { settings, spvnode, wallet } = t.context
 
   spvnode.on('tx', function (tx) {
@@ -48,6 +44,11 @@ test.serial('should sync the 20 transcations', async t => {
 
   spvnode.on('synchronized', function (newData) {
     t.log('Our node is synchronized')
+  })
+
+  wallet.on('updateFilter', async function (element) {
+    t.log(`New element to add to filter : ${element.toString('hex')}`)
+    await spvnode.updateFilter(element.toString('hex'))
   })
 
   t.log('Delete database folders')
@@ -61,6 +62,7 @@ test.serial('should sync the 20 transcations', async t => {
 
   t.log('Connect spv node to regtest node and synchronize')
   await spvnode.addPeer(settings.NODE_IP, settings.DEFAULT_PORT)
+
   await spvnode.start()
   await spvnode.synchronize()
 
@@ -68,68 +70,7 @@ test.serial('should sync the 20 transcations', async t => {
 
   await spvnode.shutdown()
 
-  t.pass()
-})
+  const balance = await wallet.getBalance()
 
-test.serial('should generate 1 more transaction', async t => {
-  const { wallet, container } = t.context
-  const containerName = (await container.inspect()).Name
-
-  // Fill our wallet with some doges
-  execSync(`docker exec ${containerName} dogecoin-cli generate 150`)
-
-  const address = await wallet.generateAddress()
-  newHash = bs58check.decode(address).slice(1).toString('hex')
-
-  t.log(`Sending to address : ${address}`)
-
-  const txid = execSync(`docker exec ${containerName} dogecoin-cli sendtoaddress ${address} 150`)
-  execSync(`docker exec ${containerName} dogecoin-cli generate 50`)
-  t.log(`Transcation ID: ${txid.toString('utf8')}`)
-
-  t.pass()
-})
-
-test.serial('should sync the 21st transaction', async t => {
-  const { settings, spvnode, wallet } = t.context
-  let counter = 0
-
-  spvnode.on('tx', function (tx) {
-    t.log(`Transaction : ${tx.id}`)
-    counter += 1
-    wallet.addTxToWallet(tx)
-  })
-
-  spvnode.on('synchronized', function (newData) {
-    t.log('Our node is synchronized')
-  })
-
-  t.log('Delete database folders')
-  fs.rmSync(path.join(settings.DATA_FOLDER, 'wallet'), { recursive: true })
-
-  t.log('Recreate folders')
-  fs.mkdirSync(path.join(settings.DATA_FOLDER, 'wallet'))
-
-  await wallet.db.openAll()
-  await wallet.init()
-
-  t.log('Connect spv node to regtest node and synchronize')
-  await spvnode.addPeer(settings.NODE_IP, settings.DEFAULT_PORT)
-  
-  // add new key to filter to all nodes
-  t.log(`new Hash : ${newHash}`)
-  await spvnode.updateFilter(newHash)
-  
-  await spvnode.start()
-  await spvnode.synchronize()
-
-  await new Promise(resolve => setTimeout(resolve, 1000))
-
-  await spvnode.shutdown()
-
-  if (counter>0) {
-    t.pass()
-  } else {
-    t.fail()
-  }
+  t.is(balance, 600000000000n)
 })
